Further improvements
SquAssets classes are now true classes and are properly extended.
A default implementation is given for `vanillaElement:getVanilla()` to warn of unimplemented methods in child classes.

Fixed a few obvious type errors in SquAssets.

The SquAPI loader no longer looks for SquAssets as the loader itself does not need it.
The individual modules will still look for it.

Greatly improved module loader to report errors that do not involve a missing module.
Also warns user if they are using the loader without any modules.

Changed event initialization to use a function instead of a global variable.
This allows users to import modules and then choose if events are started or not.
Do `module["$startEvents"]()` to start events for a specified module.
The loader will automatically run this function if the config option is enabled.